+++ src/auth/sso/mapper/SetUserValueTrait.php
@@ -82,19 +82,14 @@ trait SetUserValueTrait
 
         $field = $fieldLayout->getFieldByHandle($fieldHandle);
 
-        $field ? Craft::info(
-            sprintf(
-                "User field '%s' was found",
-                $fieldHandle
-            ),
-            'Auth'
-        ) : Craft::warning(
-            sprintf(
-                "User field '%s' was not found",
-                $fieldHandle
-            ),
-            'auth'
-        );
+        if ($field !== null) {
+            // A custom field included in this field layout
+            Craft::info(sprintf("User field '%s' was found", $fieldHandle), 'Auth');
+        } elseif (!in_array($fieldHandle, $user->attributes(), true)) {
+            // Only log if we're missing a custom field. If it's "native" properties that will always exist
+            // (firstName, lastName, fullName, email), don't log.
+            Craft::warning(sprintf("User field '%s' was not found", $fieldHandle), 'auth');
+        }
 
         return $field;
     }